pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nodi"
version = "0.1.0"
description = "Interactive data query tool for microservices architectures"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["rest", "api", "microservices", "cli", "repl", "http", "client"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Topic :: Internet :: WWW/HTTP",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "prompt-toolkit>=3.0.38",
    "pygments>=2.15.0",
    "rich>=13.0.0",
    "jsonpath-ng>=1.5.3",
    "cryptography>=41.0.0",
    "tabulate>=0.9.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
jq = [
    "pyjq>=2.6.0",
]
http2 = [
    "httpx[http2]>=0.24.0",
]
full = [
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "keyring>=24.0.0",
    "pyjq>=2.6.0",
    "httpx[http2]>=0.24.0",
]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
]
providers = [
    "pymongo>=4.0.0",
    "psycopg2-binary>=2.9.0",
    "elasticsearch>=8.0.0",
]

[project.scripts]
nodi = "nodi.__main__:main"

[project.urls]
Homepage = "https://github.com/yourusername/nodi"
Documentation = "https://github.com/yourusername/nodi/docs"
Repository = "https://github.com/yourusername/nodi"
"Bug Tracker" = "https://github.com/yourusername/nodi/issues"

[tool.setuptools]
packages = ["nodi", "nodi.config", "nodi.environment", "nodi.providers", "nodi.formatters", "nodi.plugins", "nodi.utils"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=nodi --cov-report=term-missing"
